test: split ImageCollector coverage modules
